% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_TMLE.R
\name{collapse_strata}
\alias{collapse_strata}
\title{collapse strata columns into a single strata identifier}
\usage{
collapse_strata(data, nodes)
}
\description{
collapse strata columns into a single strata identifier
}
