% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_descriptive.R
\docType{data}
\name{tmle3_Spec_descriptive}
\alias{tmle3_Spec_descriptive}
\title{Defines a tmle (minus the data)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
tmle3_Spec_descriptive
}
\description{
Current limitations:
}
\keyword{datasets}
