% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_descriptive.R
\name{tmle_descriptive}
\alias{tmle_descriptive}
\title{Risk Measures for Binary Outcomes}
\usage{
tmle_descriptive()
}
\description{
Estimates TSMs, RRs, PAR, and PAF
}
\details{
O=(W,A,Y)
W=Covariates
A=Treatment (binary or categorical)
Y=Outcome binary
}
